feat: implement hooks for commonjs and native modules and refactor the architecture
It was hard to find a good architecture on commonjs module alone. So I decided to investige native
modules as well to see how components can be shared elegatly betwwen the differents parts of the
application. This results into the old src folder being renamed as lib/client and the old lib folder
being rename as lib/server. Also, the old numberous small files of the src folder have been merged
into lib/client/es2015/script. Note that this script is separated from the rest of the client to
enable 'secret' global variables (ie global variables which are not part of the global object).

BREAKING CHANGE: The way the agent is invoked is completely changed