feat: extensive refactoring of the client to support flexible recording
The client has been refactored to support record as a library and not only as a preload. The server
is also affected to support more types of requests. For instance a single appmap can be used to
record multiple file. This is usefull to support mocha. Also the configuration as been tweaked to
use more globs.

BREAKING CHANGE: The way to launch a process-wide client has been modified. The client can now be
recorded through a library as well.