feat: post processing of event trace to recover stack structure
To bind together the trace I come with the concept of asynchronous jump. The idea is to recompose a
single callstack permutating elements of the event trace based on these jump. There exists two kinds
of jumps: manual jump and automatic jump. Manual jumps are the ones that have specific hooking:
sql-query, http-client, and http-server. The other jumps are automated and based on the async_hooks
native node module. This module is very thorough and I have to check that the automated jumps are
not taking over the regular events and cause too much overhead.

BREAKING CHANGE: The trace should be looking like a single big trace now.