fix: use posix-socket-messaging instead of fs.writeSync
Synchronously sending message to the trace server is important for two 
reasons. First, it avoid infinite loops when using asynchronous hooks. 
Second, it ensures data is sent even if the node process crash during 
the current event processing. Before we used fs.writeSync on 
net.socket.fd. But it seems to crash when writing large amount of data 
during the same event loop. This happened (EAGAIN) on running mocha on 
the juice shop. Now, we used the cleaner posix-socket module. However it 
requires node-gyp.