fix: make sure to not access 'this' in child constructors
Upon entering to a function, the agent record value of 'this'. This 
causes an error if the function is a child constructor because 'this' 
should not be accessed before calling 'super()'. Now, the empty value is 
recorded for child constructor as well as for arrows (which do not 
capture 'this').