fix: warning instead of failing upon http activity after closing
From time to time events are emitted on request/response after both have 
been closed. It might be worthwhile to investigate why it is the case. 
In the meantime, we print a warning instead of crashing the application. 
Subsequent events will not be linked to the request/response cycle 
though.