fix: support for relative paths in source map data url
Source maps can be encoded as data urls at the bottom of generated files. If such source map contains relative urls, they should be understood as urls relative to the url of the generated file rather than relative to the url of the source map (which is a data url so it does not make sense).