fix: tweak some shell commands
We should not have to explain to the user how bash work. They should now that the form `npx appmap-agent-js -- mocha test/**/*.js` means that the glob is substituted by the shell. And that if the shell does not support recursive globing they should rely on mocha's capabilities with either: `npx appmap-agent-js -- mocha --recursive test` or `npx appmap-agent-js -- mocha 'test/**/*.ts`.