fix: add sometimes missing function in classmap
The previous pruning algorithm made the false assumption that when a function is not called then none of its nested functions are called neither. This is not true for partial traces. The algorithm in this fix will investigate nested functions even if the parent function has no corresponding call event.