feat: implement fallback mechanism for synchronous communication
The prefered options is to use posix-socket to perform synchronous communication between the client and the server. Synchronous communication is important because we are monitoring asynchronous operations with the causal ordering configuration option. If the posix-socket is not available then we fallback to fs.writeSync which is expected to be slower and possible buggy for this usage.

build: fix typo on package.json