feat: use net.Socket.write rather than fs.writeSync
We need to avoid creating asynchronous resources when communicating with the server. Else, we are entering in an infinite loop when using async_hooks to perform causal ordering. Previously we used fs.writeSync to perform synchronous socket writting. But this will (probably) not work on windows plus it might have negatively affects performances. This alternative solution use the asynchronous net.Socket.write method but throttles it down. There is two options that control when net.Socket.write is actually called. First is options.heartbeat that is an interval timer to flush the internal bufer. Second is options.threshold that is the amount of messages the internal buffer can hold before flushing. Note that a too low options.threshold will still result in an infinite loop as each asynchronous resources causes several messages to be sent to the server.