feat: use more portable process spawning for configuration.command
BREAKING CHANGE: Before we considered configuration.command to be a shell script. But that does not translate well for winows. Now, we consider it is simple command that can be fed to child_process.spawn. The user will now have to provide the command as a list of token where the first token is the executable.