fix: failing assertion on identity of emit method
When using `async_hooks` the method `EventEmitter.prototype.emit` is overidden -- cf: https://github.com/nodejs/node/blob/3fb484f4e3e5944be445203a38520ef91e96de4e/lib/domain.js. This caused some assertion to failed. Rather than assuming the identity of an emitter's emit methods, I forward the emit operation on the dynamic value hold in its prototype hiearchy.

Testing this is hard because it seems to depend on the ordering of module loading. I suspect the emit mehtod is changed at the time `async_hooks` is loaded. So I add a test which use the actual application where the failure was observed. This is also good because it tests for errors on npm packaging. This test addition is accompanied with some renaming of the test files.