feat: commands are now executed via a shell rather than directly spawned
The first thing to do is add information on how to execute the command. It is in `configuration.command.options.shell`. Unlike in node, the shell is an array of tokens rather than simply the shell name. This is so the user can add some custom flags.

Commands can now longer be assumed to be provided as a list of easily parseable tokens. However, some parsing is required to add mocha hooks (it is also required for `--experimental-loader` for non-recursive process recording). Currently this parsing is not very elaborate and based on regexp. Hence, we could easily not be able to parse a mocha command. For instance `ENV_VAR=VAL npx mocha` would fail. I propose we deal with this on issue basis. Finding a shell parser is not easy. There is `https://www.npmjs.com/package/shell-quote` but it performs variable substitution. And I could not find a parser for `cmd.exe` commands.