refactor: new format marker %O which try calling String
This marker replaces the %e format which is exclusivelly for errors. The advantage of %O over %e is that it never fails. This is handy because formatting is often use for debugging so we avoid reporting error while reporting another error (harder to understand). Note that %O can trigger arbitrary code whereas %o is always safe to use.