feat: add log file as configuration option
By default everything is logged in stderr. To change this, a path can be given to `configuration.log.file`. Note that this path is accessed as a global data which is not ideal. In the future we might want to consider providing the configuration along with the dependencies to construct components.