feat: more lenient language and engine configuration field
The installer may fill the language configuration field. As it will probably not be able to determine the exact JavaScript version used, we need to allow missing version number. Missing version number will lead to make acorn parse with ecmaVersion = "latest". For consistency, we also provide the same liberty for engine and framework.