feat: use the constant javascript string for language's config field
BREAKING CHANGE: the agent will throw an error if the language 
configuration field is anything else than 'javascript'

BREAKING CHANGE: It is no longer possible to tell acorn to parse the 
code to be instrumented as a particular ecmascript version. Instead, we 
now always use the latest version. This should not cause problems 
because ecmascript is backward-compatible.