refactor: add utility component for accessing http body
Accessing on http body requires to spy on readable and writable streams. 
This is not that easy because adding event listeners may change the 
state of the stream. For instance adding the `data` listener on a 
readable stream will cause it to go in flow mode. If the observed 
application does not immediately add its own `data` listener then it 
will lose some chunks. This is why I rely on monkey patching.