refactor: change message internal representation
Related changes in the code will follow in subsequent commits. I know 
that this makes reviewing harder but the history is too messy to recover 
self-containing changes. Sorry :(

1) Before, messages (ie: trace elements) were internally represented as 
arrays instead of objects so that their stringified form would be  
shorter. But this made code harder to read. Now they are self-describing 
objects. Later, if performance is an issue, we can still add a step to 
minify their string representation but this format should only right 
before and after sending the message over the network.
2) Inline the content of `start`, `stop`, and `source` messages. There 
is no reason to introduce initialization/termination data structures.
3) Rename some event fields:
  - `index` becomes `tab`: `index` is too generic and `tab` hints to 
indentation/stack structure which is what the event trace is.
  - `type` becomes `site`: The fields of an event remain the same 
whether an event is a `before` or an `after`. Let's keep the `type` 
field for algebraic data type constructors.
  - `data` becomes `payload`: `data` is too generic.
4) Add `group` field to event. Before we relied on `before` an `after` 
async hook to group events together. But the `async_hooks` API is still 
experimental and has proven to be unreliable in the past. We now only 
rely on the `init` hook and provide each event with their current 
execution id.
5) Dedicated amend message data type. Before, amend messages were simply 
reiteration of past events. Now, they have their dedicated data type 
because it is less confusing.
6) Additional payload data types. Each event payload 
7) The message format is now completely described as a JSON schema in 
`build/schema/schema.yml`