refactor: raw message gathering in the backend
Before, the `backend` components did message sorting and passed them as 
separate arguments to the `trace` component. Now it passes a single 
trace of message. It entails that `trace/raw` returns a an array of 
message. `trace/raw` is used for testing so that we can compare traces 
with the internal format for events rather than the appmap format.