refactor: complete rewrite of event re-ordering
One of the main challenge that the js agent has to tackle is to 
rearrange trace event into something meaningful that renders well into 
the viewer. The problem comes from the ubiquitous use of execution 
events by nodejs applications which flattens their callstack. To achieve 
this we have to make some assertions which did not withstand the test of 
nextjs applications. Overall, the previous solution was messy but worked 
well for simple examples. The curent, more robust, solution is based on 
three passes of the event trace:

1) `components/trace/appmap/ordering/stack.mjs` restores a proper stack 
which may entail manufacturing jump events.
2) `components/trace/appmap/ordering/group.mjs` rearrange the stack 
based on execution id.
3) `components/trace/appmap/ordering/jump.mjs` link together before and 
after events and manufacture missing events.