feat: remove recording methods from the API
The new trace post-processing mechanism relies on a strict stack structure which can easily be broken via manual recording methods. We could change the API to make sure the stack structure is preserved but I'm not sure that anyone records event manually anyway. Let's disable this for the time being.

BREAKING CHANGES: the recording methods from the API are no longer available.