perf: shorthand for begin/end group event pair
Group-related messages compose a large part of the trace. Which is a 
problem because they carry a lot of useless information and they are 
ultimately removed from the final appmap event list. To solve this 
issue, I created a shorter representation in the frontend which is 
inflated in the backend.

[skip ci]