refactor: type convertions moved from frontend to hooks
Before, type convertion was performed in the frontend component. This 
was less than ideal because it was unclear which value provided by hooks 
had to be converted. And we do not want to perform needless convertion 
for performance reasons. Now, hooks are responsible to provide values 
that satisfies the message json schema. Serialization remains in the 
frontend component however.