refactor: use import/export in components rather than injection
This commit although very large is straightforward:

* In prod files:
  * Instead of exporting a default function that expects dependencies,
we import the proxy file residing at the top-level of each component.
  * Instead of returning the methods of the component, we export them.

* In test files:
  * Instead of loading prod files with dependency injection, we import
them.
  * Instead of loading components with dependency injection, we import
their proxy file.

The only changes in logic are located in the url and path components
because the previous system used dependency injections to test the
platform-specific parts of these components.