refactor: downgrade log of source map read failure
In order to test whether a file is excluded we first need to load its 
source map. That is because the exclusion uses the source path and not 
the generated path (if applicable). When the file is located in 
`node_modules`, the source map file is often not present. So this 
happens a lot in practice and is completely fine.