refactor: improve esm hook
- Split `lib/__common__.mjs` into `lib/loader.mjs` and 
`lib/configuration.mjs` because these are two separate concerns.
- Use global variable instead of hidden global variables. Hidden global 
variables are not part of the global object and cannot be introspected. 
That is good because they are less interfering with the observed 
application. But that is also bad because it is complex to use a dynamic 
name for these and they cannot be deleted.
- Make this global variable configurable via `configuration.hooks.esm`.