refactor: rewrite path component to match the url component
To help promote a consistent handling of paths and urls across the agent, I rewrote the path component to provide similar method to the url components. Note that as for urls, trailing (back)slashes in paths does make a difference.

Also, there is now two different files for platform-specific methods which is much easier to test.