test: add win32 drive in phony test file urls
In win32, `file:///foo/bar` is not a valid file url and will cause `url.fileURLToPath` to crash. Wherease `file:///w:/foo/bar` will succeed on both posix and win32. In most test cases, this convertion does not happen which explains why these tests did not fail. But sometimes they do, and it breaks the tests in unexpected ways. Better be consistent and use trully platform-agnostic urls.

I good idea could be to use another protocol than `file://` when convertion to path is known to not happen.