feat: default recorder is process instead of remote
If we could not detect mocha in the command we use the process recorder rather than the remote recorder.
    
For the other agents, the default recorder is the remote recorder. That is because they automatically attach themselves to processes. Hence, it makes sense to provide a on/off toggle to record appmaps. This is not the case for the js agent which require to user to explicitely prepend its command with `npx appmap-agent-js --`. In this case, requiring the user to use remote recording feels like a duplicate action and is additional friction to get appmaps to the hand of the user.