feat: preserve parsed command in configuration-accessor
Before, when the user provided the command as an array of tokens, we would in the end generate a command string from it. This simplified the implementation but requires the spawning of a shell to interpret the command string. Now we directly use the tokens to spawn the process. It is faster and provide an way for the user to bypass shell if needed.

Also the internal representation of command has been changed from `command.script` to `command.source`. This is not visiable to the user.