test: fix pg test on travis
It was a lot of trial and error to get postgres running on travis. I'm not sure it is the best way. The chmod of `/var/run/postgresql/` is ugly but it does the job.

Attempt 1: use `service: [postgres]` in `travis.yml`.
  >> Assertion failed on job for postgresql@12-main.service.

Attempt 2: use systemctl. None of the following is recognized:
  - systemctl start postgresql
  - systemctl start postgresql@13
  - systemctl start postgresql@13-main