feat: support node-fashion shell command option
Most of the time, the command should be interpreted inside a shell. This provide better cross-platform compatibility -- eg: `spawn("npm")` does not work on windows but `spawn("npm.cmd") or `spawn("npm", { shell: true})` do. So we want to execute the command inside a shell even if the user provides something like `npx appmap-agent-js -- npm run test`. Nonetheless, we provide the opportunity to bepassed the shell (as node does) via `command-options.shell = false`. The difference is that we set `command-options.shell = true` by default.

Because we rely on node to wrap the command inside a shell if needed, we can also simplify a bit the code. Now, we only need to know about the sell for escaping tokens.