feat: provide feedback if the stack overflows
Currenlty, trace post-processing relies on recursion. This code is hard to understand and will even harder to understand if we use loops. So let's provide useful feedback to the user to help reduce appmap size. Even if the stack does not overflow we still log a summary of the events at the info level.