refactor: replace expect with logError and assert
Now, internal and external error handling is only different by the error
constructor.

Before, the violation component handled external errors by abruptly
exiting the process. The rational was that user code is able to catch
errors and silence them. But I failed to realize that the user code
should not be able to silence logs. So even if the user code silences
appmap errors, there should still be a way to diagnose the issue.