refactor: use url format for code location
Inside the agent, locations are often used as strings. Lets drop the 
object format for location and always use the string format. And instead 
of using an ugly stringified json, lets use a more compact and readable 
url format: `protocol:host/path#line-column`. In the case where the base 
url already contains a hash, it is simpy overwritten. URL hashes have no 
meaning when executing javascript files so this should be safe to do.