fix: add separator for source root concatenation
Some generated source map expects that a slash be added between 
`sourceRoot` and the relative urls in `sources`. Even if it does not 
seem to be indicated in the spec, lets support this case.

> If the sources are not absolute URLs after prepending of the 
"sourceRoot", the sources are resolved relative to the SourceMap (like 
resolving script src in a html document).
-- https://sourcemaps.info/spec.html