fix: support incompatible source urls
Appmap configuration contain relative urls. But sometimes we cannot 
expressed the runtime url from the application relatively to the file 
url of the configuration file. This occurs for instance in windows where 
the application sits in `C:\` and it loads code in `D:\`. Rather than 
crashing we treat them as never matched empty files. We could 
investigate support for absolute urls in configuration file to provide 
more control to the user.

We also handle the case were the source file could not be loaded as 
empty instead of crashing. This occurs for instance with dynamically 
generated source file such as the `webpack://` url scheme.