refactor: use static imports in components
This is the first step to migrate away from parameterizing modules via
their url and using global variables and static bundling instead.

Using url search parameters to load modules was a neat idea to
parametrize modules. It relied on `./module.mjs?x=foo` having a
different cache entry than `./module.mjs?x=bar`. So a new version of the
same module was compiled for every new parameter combination which is
exactly what we wanted. For instance, this was neat to implement a
stateless logger module parametrized by the log level and the log file.
The main problem was that we had to use dynamic import to pass around
search parameters. This made static reasoning about the code harder. For
instance, bundlers could not be applied. I was aware of this limitation
before. But lately, while implementing support for Jest, I hit another
problem: Jest is executing modules in a sandbox thanks to `node:vm` and
it does not pass around search parameters. I feel this tips the scale to
rely on a less clean solution but more widely supported.