refactor: use new component system in lib files
No longer directly import components with url search parameters but use 
the loader provided in `component/load.mjs`. This loader will check 
whether a bundle with the corresponding parameter and entry point 
already exists. If not, it will write `param.mjs` and 
`components/*/index.mjs` and will bundle the entry point with `rollup`.