refactor: reorder import on top
The rationale for first reading globals was that it is kinda like we are
importing the global object. So it makes sense to put it first. But it
is purely a matter of taste. And since there exists an eslint rule to
re-order import on top, let's switch the order.