refactor: rewrite component system with regular bundling
Instead of parametrizing components with import url which made us
dynamically import modules, we use regular static bundling. The idea is
to generate files at `components/*/index.mjs` to define which instance
of a component should be bundled. The entry points of the new component
system are:
- `component/build-prod.mjs`: generate every bundles required in
production in `dist/bundles`.
- `component/build-test.mjs`: generate `components/*/*/.eslintrc` files
and route component for unit testing.
- `component/build.mjs` first apply `component/build-prod.mjs` then
`component/build-test.mjs`
- `component/lint.mjs` check that every instance of every component has
identitcal exports.