refactor: resilient access to import.meta.url
Because components are now bundled, we can no longer rely on
`import.meta.url` to refer to the exact location of the source files. At
production, it will be somewhere in `dist`. Fetching the home directory
of the agent should be done by importing `self` which only assumes that
`import.meta.url` is somewhere inside the home directory of the agent.
Which is true both in test and prod.