fix: resilient zipping of arguments and parameters
It was assumed that the number of recorded arguments is always the same 
as the number of parameters. This is actually not the case when, in the 
presence of source mapping, the original function does not have the same 
number of parameters of the generated function. For instance, this can 
be the case when the agent could not find a function at the indicated 
position and pick the next best guest. To solve this issue we simply 
truncate the zip to the smallest array.