test: data-driven system-level test
Before, system-level tests required a lot of boilerplate code. It was a
barrier to write new tests and/or modularize them. Now, each system
level-test should be an entry in `test/cases` and is essentially a mini
project. Appmaps are validated against the json schema that sits next
the them.

```
node test/system.mjs [--pack] test/cases/<name>
```

The `--pack` option indicates to pack and install the agent instead of
directly using the bin file `bin/appmap-agent-js.cjs`. This is useful to
verify that the upcoming published npm module is complete but it is
slower.

TODO: migrate the system level test suite.