feat: support additional command patterns for mocha and jest
Before, the recognition and parsing of commands was very rudimentary.
Improving this for unparsed commands required to add a basic tokenizer.
Some corner cases may not be tokenized properly. For instance `cmd.exe`
sometimes uses other escape characters than `^`. Let's see if this is an
issue in practice.

Additional support:
- Support for invocation of binary package file:
  `node node_modules/.bin/jest`
- Support for shell variable assignment:
  `VAR=VAL npx jest`
- Support for named arguments:
  `node --cpu-prof node_modules/.bin/jest