refactor: merge recorder components together
Before, we had one component per recorder. This introduce needless 
complexity because we had to merge them together to avoid introducing 
too many bundles. This merge was done in the old `recorder-cli`.

Now, we have a centralized recorder that replaces the old `recorder-cli` 
that checks for `configuration.recorder` to perform the appropriate 
actions.

NB: This change is coupled with the next commit which also merges the 
`lib/node` file of each recorder.