refactor: use module state for esm instrumentation
For consistency, we also use module state for hooking to esm loading.
Now, every hook follow the same pattern and is registered via module
state. Note that `lib/node/loader-esm.mjs` still uses a global variable
under the hood because it might be loaded twice. We remove the
possibility of `hook.esm` to be an identifier because it becomes a
no-op. I doubt anyone relied on this so let's not mark this a breaking
change.