refactor: add source map argument to agent.instrument
It is sometimes useful to explicitly pass the source map as additional 
argument to `agent.instrument`. This is the case in `jest` where 
transformers are expected to return a `{ code, map }` object. To match 
this api, we also make it possible for the content of the source map to 
already be parsed.