feat: tokenize every commands
When node is performing shell invocation of a tokenized command it 
concatenate the tokens with a simplistic space join -- cf: 
https://github.com/nodejs/node/blob/e58ed6d855e1af6579aaa50471426db8881eea99/lib/child_process.js#L619. 
Hence, if a token contains a special character, the resulting command 
can easily be messed up. I hit this for jest hooking:

```js
import { spawn } from "node:child_process";
spawn("jest", [
  "--transform",
  JSON.strinfify({
    "^": "./lib/node/jest-transformer.mjs",
  }),
], { shell: true });
```

So let's no longer rely on node's `options.shell = true` and invoke the 
shell ourselves. We can then normalize every commands into tokens. That 
simplifies the implementation because recorder hooks only have to deal 
with tokens. We hav the following cases:
- If the command is already tokenized, then `command-options.shell` 
should be `false`.
- If the command is a string and `command-options.shell` is `false`, the 
command will be tokenized similarly to `/bin/sh` even on windows.
- If the command is a string and `command-options.shell` is a string or 
`true`, the command will be wrapped inside a shell invocation. That 
means that we are no longer able to understand it and only recursive 
recording via environment hooking is available. So not support for jest 
or mocha in that case.

This is also good because we no longer have to parse shell commands 
ourselves. This was misleading for the user because the shell parsing 
was very restrictive and it was not clear which shell features was 
available. This comes at the cost of a minor breaking change because it 
is no longer possible to record jest or mocha within shell commands -- 
eg: `VAR=VAL npx jest 1> /dev/null`. Only direct invocations are 
supported -- eg: `npx jest --rootDir packages/foo`. I doubt this 
warrants a breaking change header and major version bump.