feat: programs are directly spawned by default
Now that `jest` and `mocha` can no longer be hooked when
`command-options.shell` is `true`, it seems reasonable to use `false` as
default. This should be fine for most direct commands. IMO the only
important difference is that windows sometimes requires an extension --
eg: `npx.cmd` instead of `npm` which is not the case when using a shell.
To solve this issue I introduce an additional command field:
`command-win32` which takes precedence over `command` on windows. Also,
I added an error message to make this less painful for the user.

BREAKING CHANGE: By default, programs are no longer spawned inside a
shell. Rather, they are directly spawned with a system call.