fix: instrument jest tests with transformers
Jest implements its own module system. So the regular node module
hooking does not work. Instead, we use the jest's transformer api.
Unfortunately, this option is overwritten on each new occurrence. So we
have to be careful to still apply the transformers of the user,
otherwise we might receive dialects such as react and typescript which
we are not able to instrument. Doing so requires to load the user
configuration. This represents some work because there are numerous way
to configure jest. NB: I tried to use `readConfigs` from the package
`jest-config` I could not make it work for what I needed.

The main steps of the solution are:
- Preload the jest config of the user to access its `transform` value.
- Hook to jest command by making `lib/node/transformer-jest.mjs` the
only transformer.
- This transformer receives the transformers of the user.
- It loads them and export an object that can be patched to modify any
javascript code on load.
- This object is used by `hook-module` when the recorder is `jest`.