fix: unique dynamic url for eval source
Dynamicallly evaluated code with `eval` is ... dynamic. We need to make 
sure that every invocation of `eval` is provided with a unique url to 
avoid source clash in the classmap. Introducing randomness in appmaps 
may not be ideal. If this becomes a problem in practice, we could 
investigate alternative ways to prevent source clashes. In the meantime, 
this is an easy fix.