refactor: overhaul the backend component
The three main changes are:

The backend contains multiple sessions which can be managed by two new 
message types: `open` and `close`. Before, each backend had one session 
and the `receptor` was responsible to manage sessions. This will 
simplify the changes on the `receptor` component.

Trace are no longer compiled on `stop` message because the instrumenter 
may reside on another process than the rest of the agent. Hence it is 
possible to receive the apply events before the related source. Now, the 
trace is compiled right before retrieving it with 
`compileBackendTrace(Array)`.

The backend is responsible to avoid clashes while providing the file url 
of appmaps.