fix: allow jest transformer in separate context
Jest transformer are (sometimes) executed in a different context (global 
object). So the instrumenter needs to be independent from the rest of 
the agent. The logic of `lib/node/transformer-jest.mjs` is moved to a 
dedicated component: `transformer-jest`. And, instead of using module 
state, we directly require the jest transformer component. This is more 
testable.