refactor: consistent use of backend session
Change impact of overhauling `backend` and `receptor`. The session must 
be explicitly opened and closed. It can no longer be left `null` for the 
agent to use a random value. This will help in the future when we 
develop the use case for appmapping multiple node processes at the same 
time as part of a single distributed application.