refactor: merge receptor-file and receptor-http together
Code was repeated a lot between `receptor-file` and `receptor-http`. It 
is now DRY-ed up in `receptor`. Now, an http server for the appmap 
endpoint is created on every recorder not only on remote recording. 
Also, `minifyReceptorConfiguration` is removed. It was used in `batch` 
to keep a cache of receptor. Now that we support sessions this use case 
is weak. There will be but one backend in `batch`.