fix: ignore json files in jest transformer
It was assumed that the code was always compiled by jest transformer 
into either common or native module. This is not the case, it can also 
be JSON. For now we simply ignore JSON files based on their extension. 
In the future, it might be interesting to investigate whether it is 
worth it to appmap JSON files as well.