fix: support concurrent jest test case
Jest is claiming that it is running the tests from a given file 
serially. However we detected nested calls of `beforeEach` and 
`afterEach` in one of our user. So, instead of crashing, we maintain a 
mapping of the `this` value provided to `beforeEach` and `afterEach`. 
This mapping is used to assign a track id to concurrent tests.

Unfortunately, I could not reproduce the issue so there is no associated 
test with this fix.