refactor: allow transferring multiple times the same source url
With dynamic sources, sources can share the same urls and have different 
contents. So keeping a set of sent urls in the instrumentation is no 
longer sufficient. For the time being, let's remove this check 
altogether. In the future, if this is a performance issue (which I doubt 
it will be), we could re-enable this check by hashing source contents.