refactor: no longer create empty sources
When a source content is missing or when a source url cannot be 
expressed relatively to the project directory, we no longer treat it as 
empty but simply ignore it. Source objects are directly given the 
relative url of the source rather than its absolute url and the base 
url.

NB: This commit is a bit broken because it is hard to disentangle from 
the upcoming change.