feat: support for dynamic sources
Before, apply events were linked to source files based on urls. This
forbids two sources to share the same url with two different contents.
As the agent support more transpilation tools, sources are increasingly
dynamic and this becomes more and more possible. We had the case that a
jest transformer produced two different contents for the same source
url.

To truly solve this issue we identify a source by hashing both its url
and its content. Note that if the content of a source is not loaded by
the frontend (ie source map without inlined source content) we still use
url-only identification. This should be safe in practice because these
sources are resolved by reading from the file system in the backend.