refactor: location as serializable object
Location is used to link an apply event to a function in the classmap.

Location is no longer a string but an object containing `hash`, `line`,
and `column`. This structure is now freely accessed across the codebase.
The location component only provides serialization. The reason for this
custom serialization is that it is more compact and readable than JSON.