refactor: move track logic in dedicated file
Before, the backend maintained two sets of array of messages. One for 
running tracks and one for stopped tracks which where called traces. 
This name is unfortunate because it is already used compiled array of 
messages. Instead, a boolean is added to tracks and the related logic is 
moved to a dedicated file: `components/backend/track.mjs`.

NB: The order of the messages before compilation is now chronological 
instead of being source first but that is transparent to the user.