feat: regularly flush appmaps from backend
Previously, all appmaps where kept in the backend and compile all at once at the end. This caused out of memory exceptions on large test suite. It is better remove and compile tracks that are complete from the backend at regular interval. To leverage the completion query from the backend, we add a boolean argument to `compileBackendTrack` which indicates whether it should only compile complete tracks.