refactor: bring out start message from track
This was a recent change. But I'm not satisfy by it. The idea was to only provide an array of messages to `compileTrace`. While this is appealing, it restricts what we can do in the `backend` component. For instance we would like to provide an array of `source` object to benefit from caching the result of the parser so that we do not have parse the same source over and over again every time we compile a track.

This fits in the support concurrency PR because we are going to filter messages based on their session id which entails to know the start message from the creation of the track.