refactor: no longer support sessions in the backend
Currently the backend can have multiple sessions. But this is needlessly complicated and we want to remove them. Indeed, each session is entirely independent from the others aside from the fact that they do not overwrite each-other files. In the future, independent recording can still be supported but from separate servers which is less confusing. Preventing file overwriting can still be done by specifying different output directories.

The reason why this is not listed as breaking change is because it was hidden from user in the `batch` component which never allowed more than a single open session at once.